refactor(fetch.ts, account.ts, graphql.ts): move FetchedAccount type and accountQuery from account.ts to graphql.ts
