+++ src/lib/mina/account.ts
@@ -11,11 +11,11 @@ import {
 } from '../../bindings/mina-transaction/gen/transaction.js';
 import { jsLayout } from '../../bindings/mina-transaction/gen/js-layout.js';
 import { ProvableExtended } from '../circuit-value.js';
+import { FetchedAccount } from './graphql.js';
 
-export { FetchedAccount, Account, PartialAccount };
-export { newAccount, accountQuery, parseFetchedAccount, fillPartialAccount };
+export { Account, PartialAccount };
+export { newAccount, parseFetchedAccount, fillPartialAccount };
 
-type AuthRequired = Types.Json.AuthRequired;
 type Account = Types.Account;
 const Account = Types.Account;
 
@@ -34,117 +34,31 @@ type PartialAccount = Omit<Partial<Account>, 'zkapp'> & {
   zkapp?: Partial<Account['zkapp']>;
 };
 
-// TODO auto-generate this type and the query
-type FetchedAccount = {
-  publicKey: string;
-  token: string;
-  nonce: string;
-  balance: { total: string };
-  tokenSymbol: string | null;
-  receiptChainHash: string | null;
-  timing: {
-    initialMinimumBalance: string | null;
-    cliffTime: string | null;
-    cliffAmount: string | null;
-    vestingPeriod: string | null;
-    vestingIncrement: string | null;
-  };
-  permissions: {
-    editState: AuthRequired;
-    access: AuthRequired;
-    send: AuthRequired;
-    receive: AuthRequired;
-    setDelegate: AuthRequired;
-    setPermissions: AuthRequired;
-    setVerificationKey: {
-      auth: AuthRequired;
-      txnVersion: string;
-    };
-    setZkappUri: AuthRequired;
-    editActionState: AuthRequired;
-    setTokenSymbol: AuthRequired;
-    incrementNonce: AuthRequired;
-    setVotingFor: AuthRequired;
-    setTiming: AuthRequired;
-  } | null;
-  delegateAccount: { publicKey: string } | null;
-  votingFor: string | null;
-  zkappState: string[] | null;
-  verificationKey: { verificationKey: string; hash: string } | null;
-  actionState: string[] | null;
-  provedState: boolean | null;
-  zkappUri: string | null;
-};
-const accountQuery = (publicKey: string, tokenId: string) => `{
-  account(publicKey: "${publicKey}", token: "${tokenId}") {
-    publicKey
-    token
-    nonce
-    balance { total }
-    tokenSymbol
-    receiptChainHash
-    timing {
-      initialMinimumBalance
-      cliffTime
-      cliffAmount
-      vestingPeriod
-      vestingIncrement
-    }
-    permissions {
-      editState
-      access
-      send
-      receive
-      setDelegate
-      setPermissions
-      setVerificationKey {
-        auth
-        txnVersion
-      }
-      setZkappUri
-      editActionState
-      setTokenSymbol
-      incrementNonce
-      setVotingFor
-      setTiming
-    }
-    delegateAccount { publicKey }
-    votingFor
-    zkappState
-    verificationKey {
-      verificationKey
-      hash
-    }
-    actionState
-    provedState
-    zkappUri
-  }
-}
-`;
-
 // convert FetchedAccount (from graphql) to Account (internal representation both here and in Mina)
-function parseFetchedAccount({
-  publicKey,
-  nonce,
-  zkappState,
-  balance,
-  permissions,
-  timing: {
-    cliffAmount,
-    cliffTime,
-    initialMinimumBalance,
-    vestingIncrement,
-    vestingPeriod,
-  },
-  delegateAccount,
-  receiptChainHash,
-  actionState,
-  token,
-  tokenSymbol,
-  verificationKey,
-  provedState,
-  zkappUri,
-}: FetchedAccount): Account {
+function parseFetchedAccount({ account }: FetchedAccount): Account {
+  const {
+    publicKey,
+    nonce,
+    zkappState,
+    balance,
+    permissions,
+    timing: {
+      cliffAmount,
+      cliffTime,
+      initialMinimumBalance,
+      vestingIncrement,
+      vestingPeriod,
+    },
+    delegateAccount,
+    receiptChainHash,
+    actionState,
+    token,
+    tokenSymbol,
+    verificationKey,
+    provedState,
+    zkappUri,
+  } = account;
+
   let hasZkapp =
     zkappState !== null ||
     verificationKey !== null ||
