+++ src/lib/mina/graphql.ts
@@ -1,6 +1,7 @@
 import { ActionStatesStringified, removeJsonQuotes } from '../fetch.js';
 import { UInt32 } from '../int.js';
 import { ZkappCommand } from '../account-update.js';
+import { Types } from '../../bindings/mina-transaction/types.js';
 
 export {
   type EpochData,
@@ -15,6 +16,7 @@ export {
   type ActionQueryResponse,
   type EventActionFilterOptions,
   type SendZkAppResponse,
+  type FetchedAccount,
   getEventsQuery,
   getActionsQuery,
   sendZkappQuery,
@@ -22,6 +24,52 @@ export {
   lastBlockQuery,
   lastBlockQueryFailureCheck,
   genesisConstantsQuery,
+  accountQuery,
+};
+
+type AuthRequired = Types.Json.AuthRequired;
+// TODO auto-generate this type and the query
+type FetchedAccount = {
+  account: {
+    publicKey: string;
+    token: string;
+    nonce: string;
+    balance: { total: string };
+    tokenSymbol: string | null;
+    receiptChainHash: string | null;
+    timing: {
+      initialMinimumBalance: string | null;
+      cliffTime: string | null;
+      cliffAmount: string | null;
+      vestingPeriod: string | null;
+      vestingIncrement: string | null;
+    };
+    permissions: {
+      editState: AuthRequired;
+      access: AuthRequired;
+      send: AuthRequired;
+      receive: AuthRequired;
+      setDelegate: AuthRequired;
+      setPermissions: AuthRequired;
+      setVerificationKey: {
+        auth: AuthRequired;
+        txnVersion: string;
+      };
+      setZkappUri: AuthRequired;
+      editActionState: AuthRequired;
+      setTokenSymbol: AuthRequired;
+      incrementNonce: AuthRequired;
+      setVotingFor: AuthRequired;
+      setTiming: AuthRequired;
+    } | null;
+    delegateAccount: { publicKey: string } | null;
+    votingFor: string | null;
+    zkappState: string[] | null;
+    verificationKey: { verificationKey: string; hash: string } | null;
+    actionState: string[] | null;
+    provedState: boolean | null;
+    zkappUri: string | null;
+  };
 };
 
 type GenesisConstants = {
@@ -367,3 +415,50 @@ function sendZkappQuery(json: string) {
 }
 `;
 }
+
+const accountQuery = (publicKey: string, tokenId: string) => `{
+  account(publicKey: "${publicKey}", token: "${tokenId}") {
+    publicKey
+    token
+    nonce
+    balance { total }
+    tokenSymbol
+    receiptChainHash
+    timing {
+      initialMinimumBalance
+      cliffTime
+      cliffAmount
+      vestingPeriod
+      vestingIncrement
+    }
+    permissions {
+      editState
+      access
+      send
+      receive
+      setDelegate
+      setPermissions
+      setVerificationKey {
+        auth
+        txnVersion
+      }
+      setZkappUri
+      editActionState
+      setTokenSymbol
+      incrementNonce
+      setVotingFor
+      setTiming
+    }
+    delegateAccount { publicKey }
+    votingFor
+    zkappState
+    verificationKey {
+      verificationKey
+      hash
+    }
+    actionState
+    provedState
+    zkappUri
+  }
+}
+`;
